+++ src/main/java/com/st/newtest/systemManage/Entity/UserRole.java
@@ -1,22 +1,22 @@
 package com.st.newtest.systemManage.Entity;
 
 public class UserRole {
-    private Integer uid;
-    private Integer rid;
+    private Long uid;
+    private Long rid;
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public void setUid(Integer uid) {
+    public void setUid(Long uid) {
         this.uid = uid;
     }
 
-    public Integer getRid() {
+    public Long getRid() {
         return rid;
     }
 
-    public void setRid(Integer rid) {
+    public void setRid(Long rid) {
         this.rid = rid;
     }
